requirements.txt
absl-py==2.2.2
accelerate==1.0.1
addict==2.4.0
aiofiles==23.2.1
aiohappyeyeballs==2.4.4
aiohttp==3.10.11
aiosignal==1.3.1
albumentations==0.4.3
altair==5.4.1
annotated-types==0.7.0
antlr4-python3-runtime==4.8
anyio==4.5.2
async-timeout==5.0.1
attrs==25.3.0
basicsr==1.4.2
blinker==1.8.2
Brotli @ file:///croot/brotli-split_1714483155106/work
cachetools==5.5.2
calmsize==0.1.3
certifi @ file:///home/conda/feedstock_root/build_artifacts/certifi_1725278078093/work/certifi
charset-normalizer @ file:///croot/charset-normalizer_1721748349566/work
click==8.1.8
clip==0.2.0
contourpy==1.1.1
controlnet-aux==0.0.7
cycler==0.12.1
datasets==3.1.0
diffusers==0.32.2
dill==0.3.8
docker-pycreds==0.4.0
einops==0.3.0
eval-type-backport==0.2.2
exceptiongroup==1.2.2
fairscale==0.4.13
faiss==1.7.3
fastapi==0.115.12
ffmpy==0.5.0
filelock==3.16.1
fonttools==4.57.0
frozenlist==1.5.0
fsspec==2024.9.0
ftfy==6.2.3
future==1.0.0
gensim==4.3.3
gitdb==4.0.12
GitPython==3.1.44
gmpy2 @ file:///home/conda/feedstock_root/build_artifacts/gmpy2_1715527302982/work
google-auth==2.38.0
google-auth-oauthlib==1.0.0
gradio==4.44.1
gradio-client==1.3.0
grpcio==1.70.0
h11==0.14.0
h5py==3.11.0
httpcore==1.0.8
httpx==0.28.1
huggingface-hub==0.30.1
idna @ file:///croot/idna_1714398848350/work
imageio==2.9.0
imageio-ffmpeg==0.4.2
imgaug==0.2.6
importlib-metadata==8.5.0
importlib-resources==6.4.5
invisible-watermark==0.2.0
jinja2==3.1.6
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
kagglehub==0.2.9
kiwisolver==1.4.7
kornia==0.6.0
lazy-loader==0.4
lmdb==1.6.2
loguru==0.7.3
lpips==0.1.4
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.7.5
mdurl==0.1.2
mkl-fft==1.3.1
mkl-random @ file:///tmp/build/80754af9/mkl_random_1626186064646/work
mkl-service==2.4.0
mpmath @ file:///home/conda/feedstock_root/build_artifacts/mpmath_1678228039184/work
multidict==6.1.0
multiprocess==0.70.16
narwhals==1.34.0
networkx==3.1
nltk==3.9.1
numpy==1.24.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.9.41
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
omegaconf==2.1.1
open-clip-torch==2.32.0
opencv-python==4.1.2.30
orjson==3.10.15
packaging==24.2
pandas==2.0.3
Pillow @ file:///home/conda/feedstock_root/build_artifacts/pillow_1675487152289/work
pkgutil-resolve-name==1.3.10
platformdirs==4.3.6
propcache==0.2.0
protobuf==5.29.4
psutil==7.0.0
pudb==2019.2
pyarrow==17.0.0
pyasn1==0.6.1
pyasn1-modules==0.4.2
pydantic==2.10.6
pydantic-core==2.27.2
pydeck==0.9.1
pyDeprecate==0.3.1
pydub==0.25.1
pygments==2.19.1
pyparsing==3.1.4
PySocks @ file:///tmp/build/80754af9/pysocks_1605305779399/work
python-dateutil==2.9.0.post0
python-multipart==0.0.20
pytorch-lightning==1.4.2
pytorch-memlab==0.3.0
pytz==2025.2
PyWavelets==1.4.1
PyYAML==6.0.2
referencing==0.35.1
regex==2024.11.6
requests @ file:///croot/requests_1721410876868/work
requests-oauthlib==2.0.0
rich==13.9.4
rpds-py==0.20.1
rsa==4.9
ruff==0.11.6
sacremoses==0.1.1
safetensors==0.5.3
scikit-image==0.20.0
scikit-learn==1.3.2
scipy==1.9.1
seaborn==0.13.2
semantic-version==2.10.0
sentry-sdk==2.29.1
setproctitle==1.3.6
shellingham==1.5.4
six @ file:///tmp/build/80754af9/six_1644875935023/work
smart-open==7.1.0
smmap==5.0.2
sniffio==1.3.1
starlette==0.44.0
streamlit==1.40.1
sympy @ file:///home/conda/feedstock_root/build_artifacts/sympy_1728484478345/work
taming-transformers==0.0.1
tb-nightly==2.14.0a20230808
tenacity==9.0.0
tensorboard==2.14.0
tensorboard-data-server==0.7.2
test-tube==0.7.5
thop==0.1.1.post2209072238
threadpoolctl==3.5.0
tifffile==2023.7.10
timm==1.0.15
tokenizers==0.15.2
toml==0.10.2
tomli==2.2.1
tomlkit==0.12.0
torch @ file:///hdd/zh/Env/torch-2.2.2%2Bcu121-cp38-cp38-linux_x86_64.whl
torch-fidelity==0.3.0
torchaudio @ file:///hdd/zh/Env/torchaudio-2.2.2%2Bcu121-cp38-cp38-linux_x86_64.whl
torchmetrics==0.6.0
torchstat==0.0.7
torchvision @ file:///hdd/zh/Env/torchvision-0.17.2%2Bcu121-cp38-cp38-linux_x86_64.whl
tornado==6.4.2
tqdm==4.67.1
transformers==4.36.2
triton==2.2.0
typer==0.15.2
typing-extensions==4.13.2
tzdata==2025.2
urllib3 @ file:///croot/urllib3_1727769808118/work
urwid==2.6.16
uvicorn==0.33.0
wandb==0.19.11
watchdog==4.0.2
wcwidth==0.2.13
websockets==12.0
werkzeug==3.0.6
wrapt==1.17.2
xxhash==3.5.0
yapf==0.43.0
yarl==1.15.2
zipp==3.20.2
